move code to edit dir
