+++ src/app/activities/[activityId]/edit/components/ImageUpload.js
