+++ src/app/activities/[activityId]/edit/page.js
@@ -2,8 +2,8 @@ import ImagesMultipleUpload from "./components/ImageMultipleUpload";
 import ActivitiesSection from "./sections/ActivitiesSection";
 import SecondaryImagesSection from "./sections/SecondaryImagesSection";
 import TermsAndConditionsSection from "./sections/TermsAndConditionsSection";
-import {MapSection} from "@/app/activities/[activityId]/sections/MapSection";
-import {SubmitSection} from "@/app/activities/[activityId]/sections/SubmitSection";
+import {MapSection} from "@/app/activities/[activityId]/edit/sections/MapSection";
+import {SubmitSection} from "@/app/activities/[activityId]/edit/sections/SubmitSection";
 
 export default function ActivityDetailsPage() {
   return (
