+++ src/app/activities/[activityId]/edit/sections/SecondaryImagesSection.js
