+++ src/app/activities/[activityId]/edit/sections/SubmitSection.js
